+++ src/lib/Resolver.ts
@@ -1,0 +1,4 @@
+import { Question } from './dns/Question';
+import { Message } from './dns/Message';
+
+export type Resolver = (question: Question) => Promise<Message | Buffer>;
