+++ src/lib/VerificationOptions.ts
@@ -1,5 +1,5 @@
 import { DatePeriod } from './DatePeriod';
-import { DsData } from '../rdata/DsData';
+import { DsData } from './rdata/DsData';
 
 export interface VerificationOptions {
   readonly dateOrPeriod: Date | DatePeriod;
