--- src/lib/verification/Resolver.ts
@@ -1,4 +1,0 @@
-import { Question } from '../dns/Question';
-import { Message } from '../dns/Message';
-
-export type Resolver = (question: Question) => Promise<Message | Buffer>;
